[Fixes #4310] Moved AuthorizeFilter and CorsAuthorizationFilter to a public namespace
