--- src/Microsoft.AspNetCore.Mvc.Abstractions/Filters/IAllowAnonymousFilter.cs
@@ -1,9 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNetCore.Mvc.Filters
-{
-    public interface IAllowAnonymousFilter : IFilterMetadata
-    {
-    }
-}
