+++ src/Microsoft.AspNetCore.Mvc.Core/Authorization/AllowAnonymousFilter.cs
@@ -1,0 +1,12 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Mvc.Authorization
+{
+    /// <summary>
+    /// An implementation of <see cref="IAllowAnonymousFilter"/>
+    /// </summary>
+    public class AllowAnonymousFilter : IAllowAnonymousFilter
+    {
+    }
+}
